refactor(session): move the multi-client registry into internal/session/clientreg
internal/session bundled three unrelated subsystems. The multi-client registry
— client IDs, per-buffer cursors, LRU eviction, idle reaping — is the most
clearly separate: the reference graph shows exactly one edge between it and the
rest of the package (client_state.go reading clientIDLength from
client_registry.go), i.e. zero edges leaving the cluster.

The package doc was the tell. doc.go described internal/session as
"multi-client session management" and mentioned nothing about snapshots or
diffing, even though snapshot capture/compare is the bulk of the package and
the only part wired into an MCP tool. The doc had been describing this cluster
alone. Both docs are now rewritten to match what each package actually holds.

  internal/session            20 -> 18 files
  internal/session/clientreg   2 files (new)

No behavior change: same registry semantics, same exported method set, tests
moved with the code (clientreg 97.3% covered).
