+++ internal/session/doc.go
@@ -1,21 +1,20 @@
 // Purpose: Implement doc.go internal behavior used by MCP runtime features.
 // Why: Maintains stable server behavior across tool and transport paths.
-// Docs: docs/features/feature/pagination/index.md
+// Docs: docs/features/feature/request-session-correlation/index.md
 
-// doc.go — Package documentation for multi-client session management.
+// doc.go — Package documentation for named session snapshots.
 
-// Package session provides multi-client session management for the Kaboom MCP server.
+// Package session manages named snapshots of browser state and the diff_sessions
+// MCP tool that captures, lists, compares and deletes them.
 //
-// Features:
-//   - Client registration and isolation (each client has unique ID and working directory)
-//   - Session state tracking (active, idle, stale detection)
-//   - Last poll timestamp tracking for client health monitoring
-//   - Thread-safe client registry with automatic cleanup
-//   - Verification loop support for test flake detection
+// The SessionManager reads live state through a CaptureStateReader, stores
+// bounded named snapshots with insertion-order eviction, and answers Compare by
+// delegating to the pure diff engine in the snapdiff subpackage.
 //
-// The ClientRegistry maintains state for multiple Claude Code sessions connecting
-// to a single Kaboom server instance. Each client is identified by X-Kaboom-Client
-// header and maintains isolated state (current working directory, last poll time, etc.).
+// Two neighbours that used to live in this package now have their own:
 //
-// Clients are considered stale if they haven't polled within 3 seconds.
+//   - clientreg: multi-client registry (client IDs, cursors, LRU, idle reaping).
+//   - verify: the verify_fix before/after verification loop.
+//
+// Both are independent of snapshot storage; verify shares only CaptureStateReader.
 package session
